公共win配置
