--- src/views/index/frame/login/app.html
@@ -1,12 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="maximum-scale=1.0, minimum-scale=1.0, user-scalable=0, initial-scale=1.0, width=device-width" />
-    <meta name="format-detection" content="telephone=no, email=no, date=no, address=no">
-    <title></title>
-</head>
-<body>
-    <div id='app'></div>
-</body>
-</html>
