+++ src/views/index/login/app.js
@@ -4,16 +4,16 @@ import 'muse-ui/lib/styles/base.less';
 // import 'muse-ui/lib/styles/theme.less';
 import 'assets/css/global.less';
 import 'assets/fonts/iconfont.css';
-
+var vm;
 if (process.env === 'production') {
   window.apiready = function () {
-    new Vue({
+    vm=new Vue({
       el: '#app',
       render: h => h(App)
     });
   };
 } else {
-  new Vue({
+  vm=new Vue({
     el: '#app',
     render: h => h(App)
   });
