+++ src/views/index/win/app.vue
@@ -5,11 +5,36 @@
       :z-depth="1"
       class="header"
       :title="wtitle">
+      <Button
+        icon
+        slot="left"
+        v-if="hasLeft"
+        @click="clickLeft"
+      >
+        <Icon
+          value=":icon-left"
+          color="#666"
+          :size="24"
+        />
+      </Button>
+      <Button
+        icon
+        slot="right"
+        v-if="hasRight"
+        @click="clickRight"
+      >
+        <Icon
+          value=":icon-kefunv"
+          color="#666"
+          :size="24"
+        />
+      </Button>
     </UIHeader>
 </template>
 
 <script>
 import tool from 'util/tools';
+import { Icon, Button } from 'muse-ui';
 import UIHeader from 'components/UIHeader';
 
 export default {
@@ -18,13 +43,26 @@ export default {
         wtitle:window.api.pageParam.wtitle,
         fname:window.api.pageParam.fname,
         furl:window.api.pageParam.furl,
+        hasLeft:window.api.pageParam.hasLeft,
+        hasRight:window.api.pageParam.hasRight,
     };
   },
   components: {
-   UIHeader
+      Button,
+      Icon,
+      UIHeader
   },
   methods: {
-    
+    clickLeft(){
+        window.api.closeWin();
+        // window.api.execScript({
+        //     frameName: this.fname,
+        //     script:'window.clickLeft();'
+        // });
+    },
+    clickRight(){
+        alert("right");
+    }
   },
   mounted () {
       var headerH=tool.fixStatusBar(tool.dom(".header"));
